reorganize app in prep for dynamic imports. in progess. #10.
