+++ src/components/Visualization/habitat_risk_assessment/Plot/Legend.js
