+++ src/components/Visualization/habitat_risk_assessment/reducers/index.js
