+++ src/components/Visualization/habitat_risk_assessment/reducers/reducer_file_suffix.js
