+++ src/index.html
@@ -18,11 +18,11 @@
   />
   <link
     rel="stylesheet"
-    href="./components/Map/style.css"
+    href="./components/Visualization/habitat_risk_assessment/Map/style.css"
   />
   <link
     rel="stylesheet"
-    href="./components/Plot/style.css"
+    href="./components/Visualization/habitat_risk_assessment/Plot/style.css"
   />
 </head>
 <body style="background-color: rgba(0,0,0,0); margin: 0; padding: 20px" >
