+++ src/valid_HRA_args.js
@@ -1,7 +1,7 @@
 const SAMPLE_DATA = 'C:/Users/dmf/projects/invest/data/invest-sample-data/HabitatRiskAssess/Input/'
 
 const MODULE_NAME = 'natcap.invest.hra'
-const MODEL_NAME = 'carbon'
+const MODEL_NAME = 'habitat_risk_assessment'
 const MODEL_DOCS = 'C:/InVEST_3.7.0_x86/documentation/userguide/habitat_risk_assessment.html'
 
 const MODEL_ARGS = {
